+++ pgpack/common/__init__.py
@@ -1,6 +1,4 @@
-from .cast_dataframes import (
-    pandas_astype,
-)
+from .cast_dataframes import pandas_astype
 from .detector import detect_oid
 from .errors import (
     PGPackError,
@@ -8,7 +6,10 @@ from .errors import (
     PGPackModeError,
     PGPackMetadataCrcError,
 )
-from .header import HEADER
+from .headers import (
+    HEADER,
+    S3_FILE_HEADER,
+)
 from .metadata import (
     metadata_from_frame,
     metadata_reader,
@@ -17,14 +18,15 @@ from .param import PGParam
 
 
 __all__ = (
-    "HEADER",
-    "detect_oid",
-    "metadata_from_frame",
-    "metadata_reader",
-    "pandas_astype",
     "PGPackError",
     "PGPackHeaderError",
     "PGPackModeError",
     "PGPackMetadataCrcError",
     "PGParam",
+    "detect_oid",
+    "metadata_from_frame",
+    "metadata_reader",
+    "pandas_astype",
+    "HEADER",
+    "S3_FILE_HEADER",
 )