+++ pgpack/common/headers.py
@@ -1,0 +1,2 @@
+HEADER = b"PGPACK\n\x00"
+S3_FILE_HEADER = b"pgpack_s3_object"